requirements.txt
django
mechanize
beautifulsoup4
lxml
django-localflavor
pytz